arch/arm64: Use variables of the right size for msr/mrs opcodes
They do 64bit accesses, and gcc does the necessary fix ups to handle
32bit values as zero-padded 64bit values.
clang, however, isn't happy with it.

Reviewed-on: https://review.coreboot.org/19661
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/arch/arm64/armv8/lib/clock.c
@@ -23,5 +23,5 @@
 
 void set_cntfrq(uint32_t freq)
 {
-	__asm__ __volatile__("msr cntfrq_el0, %0" :: "r"(freq));
+	__asm__ __volatile__("msr cntfrq_el0, %0" :: "r"((uint64_t)freq));
 }
